Inspired by the 3D alpha shape, I rewrote the code to get the Alpha_shape_2 working with the Periodic_2_triangulation_2.
The main change to master is that the computation now uses:
      return _gt.compute_squared_radius_2_object()(this->point(f,0), this->point(f,1), this->point(f,2));
This will always give inexact results since the points are constructed first and then given to the Compute_squared_radius_2 object.
Note that we use the method from _gt since the default constructor does not exist for the periodic triangulation.